Move Detector and Images Generator to package directory
