+++ maplap/Detector.py
